requirements/development.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/development.in
#
ansicon==1.89.0
    # via jinxed
anyascii==0.3.2
    # via wagtail
asgiref==3.8.1
    # via django
bcrypt==4.2.0
    # via paramiko
beautifulsoup4==4.11.2
    # via
    #   wagtail
    #   webtest
black==24.8.0
    # via
    #   -r c:\KH\klimaat-helpdesk\requirements\testing.in
    #   flake8-black
blessed==1.20.0
    # via inquirer
boolean-py==4.0
    # via license-expression
brotli==1.1.0
    # via whitenoise
cachecontrol[filecache]==0.14.0
    # via
    #   cachecontrol
    #   pip-audit
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage[toml]==7.6.1
    # via pytest-cov
cryptography==43.0.1
    # via paramiko
cyclonedx-python-lib==7.6.0
    # via pip-audit
decorator==5.1.1
    # via fabric
defusedxml==0.7.1
    # via
    #   py-serializable
    #   willow
deprecated==1.2.14
    # via fabric
dj-database-url==2.3.0
    # via -r requirements/development.in
django==5.2
    # via
    #   -r c:\KH\klimaat-helpdesk\requirements\base.in
    #   dj-database-url
    #   django-filter
    #   django-health-check
    #   django-modelcluster
    #   django-permissionedforms
    #   django-stubs-ext
    #   django-taggit
    #   django-tasks
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
django-filter==23.5
    # via wagtail
django-health-check==3.18.3
    # via -r c:\KH\klimaat-helpdesk\requirements\base.in
django-modelcluster==6.3
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-stubs-ext==5.2.2
    # via django-tasks
django-taggit==6.1.0
    # via wagtail
django-tasks==0.8.1
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
django-webtest==1.9.12
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
djangorestframework==3.15.2
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
editor==1.6.6
    # via inquirer
et-xmlfile==1.1.0
    # via openpyxl
fabric==3.2.2
    # via -r requirements/development.in
factory-boy==3.3.1
    # via wagtail-factories
faker==28.4.1
    # via factory-boy
filelock==3.15.4
    # via cachecontrol
filetype==1.2.0
    # via willow
flake8==7.1.1
    # via
    #   -r c:\KH\klimaat-helpdesk\requirements\testing.in
    #   flake8-black
    #   flake8-debugger
    #   flake8-isort
flake8-black==0.3.6
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
flake8-debugger==4.1.2
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
flake8-isort==6.1.1
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r requirements/development.in
gunicorn==23.0.0
    # via -r requirements/development.in
h11==0.16.0
    # via uvicorn
html5lib==1.1
    # via pip-audit
idna==3.8
    # via requests
iniconfig==2.0.0
    # via pytest
inquirer==3.4.0
    # via -r requirements/development.in
invoke==2.2.0
    # via fabric
isort==5.13.2
    # via
    #   -r c:\KH\klimaat-helpdesk\requirements\testing.in
    #   flake8-isort
jinxed==1.3.0
    # via blessed
laces==0.1.2
    # via wagtail
license-expression==30.3.1
    # via cyclonedx-python-lib
markdown-it-py==3.0.0
    # via rich
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.0.8
    # via cachecontrol
mypy-extensions==1.0.0
    # via black
openpyxl==3.1.5
    # via wagtail
packageurl-python==0.15.6
    # via cyclonedx-python-lib
packaging==24.1
    # via
    #   black
    #   gunicorn
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
paramiko==3.4.1
    # via fabric
pathspec==0.12.1
    # via black
pep8==1.7.1
    # via -r requirements/development.in
pillow==10.4.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.18.0
    # via willow
pip-api==0.0.34
    # via pip-audit
pip-audit==2.7.3
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
pip-requirements-parser==32.0.1
    # via pip-audit
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements/development.in
py-serializable==1.1.0
    # via cyclonedx-python-lib
pycodestyle==2.12.1
    # via
    #   flake8
    #   flake8-debugger
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pynacl==1.5.0
    # via paramiko
pyparsing==3.1.4
    # via pip-requirements-parser
pytest==8.3.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
pytest-django==4.9.0
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
python-dateutil==2.9.0.post0
    # via faker
python-gitlab==4.10.0
    # via -r c:\KH\klimaat-helpdesk\requirements\base.in
pytz==2024.1
    # via django-modelcluster
pyyaml==6.0.2
    # via responses
readchar==4.2.0
    # via inquirer
requests==2.32.3
    # via
    #   cachecontrol
    #   pip-audit
    #   python-gitlab
    #   requests-toolbelt
    #   responses
    #   wagtail
requests-toolbelt==1.0.0
    # via python-gitlab
responses==0.25.3
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
rich==13.8.0
    # via pip-audit
runs==1.2.2
    # via editor
six==1.16.0
    # via
    #   blessed
    #   html5lib
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.1
    # via django
telepath==0.3.1
    # via wagtail
termcolor==2.4.0
    # via -r requirements/development.in
toml==0.10.2
    # via pip-audit
typing-extensions==4.13.2
    # via
    #   dj-database-url
    #   django-stubs-ext
    #   django-tasks
tzdata==2025.2
    # via django
urllib3==2.2.2
    # via
    #   requests
    #   responses
uvicorn==0.34.2
    # via -r requirements/development.in
wagtail==7.1
    # via
    #   -r c:\KH\klimaat-helpdesk\requirements\base.in
    #   wagtail-factories
wagtail-factories==4.2.1
    # via -r c:\KH\klimaat-helpdesk\requirements\testing.in
waitress==3.0.0
    # via webtest
wcwidth==0.2.13
    # via blessed
webencodings==0.5.1
    # via html5lib
webob==1.8.8
    # via webtest
webtest==3.0.1
    # via django-webtest
whitenoise[brotli]==6.9.0
    # via -r requirements/development.in
willow[heif]==1.11.0
    # via
    #   wagtail
    #   willow
wrapt==1.16.0
    # via deprecated
xmod==1.8.1
    # via
    #   editor
    #   runs

# The following packages are considered to be unsafe in a requirements file:
# pip
